feat: migration controller to TS
